+++ src/urban/odx_matrix.hpp
@@ -10,29 +10,29 @@
 namespace urban {
 
 	class odx_pair_info {
-		int _morning_rush_hour;
-		int _midday;
-		int _afternoon_rush_hour;
-		int _happy_hour;
-		int _night_time;
-		int _total;
+		float _morning_rush_hour;
+		float _midday;
+		float _afternoon_rush_hour;
+		float _happy_hour;
+		float _night_time;
+		float _total;
 
 		public:
 		odx_pair_info(
-			int morning_rush_hour,
-			int midday,
-			int afternoon_rush_hour,
-			int happy_hour,
-			int night_time,
-			int total
+			float morning_rush_hour,
+			float midday,
+			float afternoon_rush_hour,
+			float happy_hour,
+			float night_time,
+			float total
 		);
 		odx_pair_info();
-		int get_morning_rush_hour() const;
-		int get_midday() const;
-		int get_afternoon_rush_hour() const;
-		int get_happy_hour() const;
-		int get_night_time() const;
-		int get_total() const;
+		float get_morning_rush_hour() const;
+		float get_midday() const;
+		float get_afternoon_rush_hour() const;
+		float get_happy_hour() const;
+		float get_night_time() const;
+		float get_total() const;
 	};
 
 	class odx_entry {
@@ -40,12 +40,12 @@ namespace urban {
 		
 		public:
 		odx_entry();
-		int get_morning_rush_hour(std::pair<int, int> destin);
-		int get_midday(std::pair<int, int> destin);
-		int get_afternoon_rush_hour(std::pair<int, int> destin);
-		int get_happy_hour(std::pair<int, int> destin);
-		int get_night_time(std::pair<int, int> destin);
-		int get_total(std::pair<int, int> destin);
+		float get_morning_rush_hour(std::pair<int, int> destin);
+		float get_midday(std::pair<int, int> destin);
+		float get_afternoon_rush_hour(std::pair<int, int> destin);
+		float get_happy_hour(std::pair<int, int> destin);
+		float get_night_time(std::pair<int, int> destin);
+		float get_total(std::pair<int, int> destin);
 		void add_odx_pair_info(
 			std::pair<int, int> destin, 
 			odx_pair_info info
@@ -64,13 +64,13 @@ namespace urban {
 
 		public:
 		static odx_matrix* instance();
-		int get_morning_rush_hour(std::pair<int, int> origin, std::pair<int, int> destin);
-		int get_midday(std::pair<int, int> origin, std::pair<int, int> destin);
-		int get_afternoon_rush_hour(std::pair<int, int> origin, std::pair<int, int> destin);
-		int get_happy_hour(std::pair<int, int> origin, std::pair<int, int> destin);
-		int get_night_time(std::pair<int, int> origin, std::pair<int, int> destin);
-		int get_total(std::pair<int, int> origin, std::pair<int, int> destin);
-		int get_in_day_time(std::string day_time, std::pair<int, int> origin, std::pair<int, int> destin);
+		float get_morning_rush_hour(std::pair<int, int> origin, std::pair<int, int> destin);
+		float get_midday(std::pair<int, int> origin, std::pair<int, int> destin);
+		float get_afternoon_rush_hour(std::pair<int, int> origin, std::pair<int, int> destin);
+		float get_happy_hour(std::pair<int, int> origin, std::pair<int, int> destin);
+		float get_night_time(std::pair<int, int> origin, std::pair<int, int> destin);
+		float get_total(std::pair<int, int> origin, std::pair<int, int> destin);
+		float get_in_day_time(std::string day_time, std::pair<int, int> origin, std::pair<int, int> destin);
 		void add_odx_pair_info(
 			std::pair<int, int> origin, 
 			std::pair<int, int> destin,
